+++ Mora/CHANGELOG.md
@@ -2,6 +2,10 @@
 - Light/dark: allow multiple blocks for light/dark with user-specified start + end
 - if user exits QInputDialog that opens from score_data(self), do not score the data!
 
+## 2025-01-20
+### Fixed
+- Allow multiple light/dark periods to be applied to a single recording.
+
 ## 2025-01-07
 ### Fixed
 - Rule based filter now applied when scoring data in `update_objects.py`